improve tests
